+++ src/dotnet.ts
@@ -1,5 +1,6 @@
+export * as msbuildProject from "./dotnet/MSBuildProject.js"
+export * as msbuildProjectProperties from "./dotnet/MSBuildProjectProperties.js"
+export * as createDummyNupkg from './dotnet/createDummyNupkg.js'
 export * as dotnetGHPR from "./dotnet/dotnetGHPR.js"
 export * as dotnetGLPR from "./dotnet/dotnetGLPR.js"
 export * as dotnetHelpers from "./dotnet/dotnetHelpers.js"
-export * as msbuildProject from "./dotnet/MSBuildProject.js"
-export * as msbuildProjectProperties from "./dotnet/MSBuildProjectProperties.js"