+++ src/dotnet/dotnetHelpers.ts
@@ -1,8 +1,8 @@
 import { ok } from 'node:assert/strict';
+import { getEnvVarValue } from '../envUtils.js';
+import { MSBuildProject, MSBuildProjectPreDefinedProperties } from './MSBuildProject.js';
 import { getGithubNugetRegistryPair, nugetGitHubUrlBase } from './dotnetGHPR.js';
 import { getGitlabNugetRegistryPair } from './dotnetGLPR.js';
-import { MSBuildProject, MSBuildProjectPreDefinedProperties } from './MSBuildProject.js';
-
 
 function formatDotnetPublish(projectsToPublish: string[], publishProperties: string[]): string {
 	/* Fun Fact: You can define a property and get the evaluated value in the same command!
@@ -175,9 +175,9 @@ export async function configureDotnetNugetPush(
 	return registries
 		.map(
 			(registry) => {
-				const tokenValue = process.env[registry.tokenEnvVar];
-				ok(tokenValue, `The environment variable ${registry.tokenEnvVar} is undefined!`);
-				`dotnet nuget push ${nupkgDir} --source ${registry.url} --token ${tokenValue}`
+				const tokenValue = getEnvVarValue(registry.tokenEnvVar);
+				ok(getEnvVarValue('SKIP_TOKEN') === 'true' || tokenValue, `The environment variable ${registry.tokenEnvVar} is undefined!`);
+				`dotnet nuget push ${nupkgDir} --source ${registry.url} --token ${tokenValue ?? '**placeholder**'}`
 			}
 		)
 		.join(' && ');