+++ tests/configDotnet.test.ts
@@ -1,6 +1,9 @@
-import { describe, it } from 'node:test';
 import { appendPlugins, getConfig, insertAndEditPlugins } from '@halospv3/hce.shared-config/semanticReleaseConfigDotnet';
 import { notDeepStrictEqual, ok, strictEqual } from 'node:assert';
+import { unlinkSync, writeFileSync } from 'node:fs';
+import { env } from 'node:process';
+import { describe, it } from 'node:test';
+import { fileSync, setGracefulCleanup } from 'tmp';
 
 await describe('configDotnet', async () => {
 	await describe('appendPlugins', () => {
@@ -33,7 +36,32 @@ await describe('configDotnet', async () => {
 	await describe('getConfig', async () => {
 		strictEqual(getConfig.name, 'getConfig');
 
-		await it('throws Error when projectsToPublish is an empty array, projectsToPackAndPush are both empty arrays.', () => {
+
+		await it('does not throw when projectToPackAndPush contains at least one item', () => {
+			env['SKIP_TOKEN'] = 'true';
+			const actual = (() => {
+				setGracefulCleanup();
+				const tmpProj = fileSync({ postfix: '.csproj', discardDescriptor: true });
+				writeFileSync(tmpProj.name, '<Project> <PropertyGroup> <TargetFramework>net6.0</TargetFramework> <RuntimeIdentifier>win7-x86</RuntimeIdentifier> </PropertyGroup> </Project>')
+				try {
+					return getConfig([tmpProj.name], [tmpProj.name])
+				}
+				catch (err) {
+					if (err instanceof AggregateError)
+						return err;
+					else if (err instanceof Error)
+						return err;
+					else
+						return new Error(String(err));
+				}
+				finally {
+					unlinkSync(tmpProj.name)
+				}
+			})();
+			ok(!(actual instanceof Error), '`actual` should not be an Error.\n' + actual.stack)
+		})
+
+		await it('throws Error when projectsToPublish is an empty array.', () => {
 			const actual = (() => {
 				try {
 					return getConfig([], false);
@@ -42,7 +70,7 @@ await describe('configDotnet', async () => {
 				}
 			})();
 			ok(actual instanceof Error)
-			strictEqual(actual.message, "getConfig cannot continue. One or more errors occurred.\nprojectsToPublish.length must be > 0 or PROJECTS_TO_PUBLISH must be defined and contain at least one path.")
+			ok(actual.message.includes("projectsToPublish.length must be > 0 or PROJECTS_TO_PUBLISH must be defined and contain at least one path."));
 		})
 	});
 });