+++ tests/dotnet/MSBuildProjectProperties.test.ts
@@ -1,4 +1,9 @@
-import { todo } from 'node:test';
 import { MSBuildProjectProperties } from '@halospv3/hce.shared-config/dotnet/MSBuildProjectProperties';
+import { ok } from 'node:assert/strict';
+import { describe, it } from 'node:test';
 
-await todo(MSBuildProjectProperties.name);
+await describe('MSBuildProjectProperties', async () => {
+    await it('exists', () => {
+        ok(MSBuildProjectProperties);
+    })
+});